Move ShardingSphere statistics to kernel schedule module
