+++ kernel/schedule/core/src/test/java/org/apache/shardingsphere/schedule/core/job/statistics/StatisticsScheduleCollectorTest.java
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.data.pipeline.core.listener;
+package org.apache.shardingsphere.schedule.core.job.statistics;
 
-import org.apache.shardingsphere.data.pipeline.core.listener.ShardingSphereStatisticsScheduleCollector.ShardingSphereDataCollectorRunnable;
 import org.apache.shardingsphere.infra.config.props.ConfigurationProperties;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
@@ -29,6 +28,7 @@ import org.apache.shardingsphere.infra.metadata.statistics.ShardingSphereSchemaD
 import org.apache.shardingsphere.infra.metadata.statistics.ShardingSphereStatistics;
 import org.apache.shardingsphere.infra.metadata.statistics.ShardingSphereTableData;
 import org.apache.shardingsphere.mode.manager.ContextManager;
+import org.apache.shardingsphere.schedule.core.job.statistics.StatisticsScheduleCollector.StatisticsCollectorRunnable;
 import org.junit.jupiter.api.Test;
 
 import java.sql.Types;
@@ -42,7 +42,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class ShardingSphereDataCollectorTest {
+class StatisticsScheduleCollectorTest {
     
     @Test
     void assertCollect() {
@@ -52,7 +52,7 @@ class ShardingSphereDataCollectorTest {
         ShardingSphereMetaData metaData = mockMetaData();
         when(contextManager.getMetaDataContexts().getMetaData()).thenReturn(metaData);
         when(contextManager.getMetaDataContexts().getMetaData().getProps()).thenReturn(new ConfigurationProperties(new Properties()));
-        new ShardingSphereDataCollectorRunnable(contextManager).run();
+        new StatisticsCollectorRunnable(contextManager).run();
         verify(contextManager).getInstanceContext();
     }
