+++ kernel/schedule/pom.xml
@@ -20,26 +20,14 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.shardingsphere</groupId>
-        <artifactId>shardingsphere</artifactId>
+        <artifactId>shardingsphere-kernel</artifactId>
         <version>5.4.2-SNAPSHOT</version>
     </parent>
-    <artifactId>shardingsphere-kernel</artifactId>
+    <artifactId>shardingsphere-schedule</artifactId>
     <packaging>pom</packaging>
     <name>${project.artifactId}</name>
     
     <modules>
-        <module>single</module>
-        <module>authority</module>
-        <module>transaction</module>
-        <module>time-service</module>
-        <module>data-pipeline</module>
-        <module>sql-federation</module>
-        <module>sql-parser</module>
-        <module>sql-translator</module>
-        <module>traffic</module>
-        <module>logging</module>
-        <module>metadata</module>
-        <module>global-clock</module>
-        <module>schedule</module>
+        <module>core</module>
     </modules>
 </project>
